Added Filters, Search & Order.
In this commit, new methods for filtering, searching and ordering are added. Also, Internal working of GetColumn & GetAllRows method are changed which results faster data loading for more than 200 rows. Later, it looks to Error to finishi fetching data but now, it will directly understand on which point it would have to finish fetching of data from given size.